cofflyze-model/requirements.txt
Flask==2.3.3
tensorflow==2.17.0
google-cloud-storage==2.11.0
mysql-connector-python==8.2.0
gunicorn==21.2.0
python-dotenv
google-cloud-firestore
pillow
flask
pytz